+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - The command line options are more consistent now
 - Added a validation command for submission files
 - A user can now add a file to a submission
+- The system enforces now the file extension .zip when downloading
+  submission or dataset files
 
 
 ## CHANGES in v0.2: